LayoutTests/editing/pasteboard/copy-content-with-display-none-exiting-ancestors-expected.txt

| <i>
|   "hello "
| <b>
|   "world"
|   " "
| "bar<#selection-caret>"
